requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
langgraph>=0.2.0
langchain-google-genai>=2.0.0
google-generativeai>=0.8.0
apscheduler>=3.10.4
httpx>=0.27.0
python-dotenv>=1.0.0
PyGithub>=2.3.0
chromadb>=0.5.0
sqlalchemy>=2.0.0
pydantic>=2.8.0
rich>=13.7.0
tweepy>=4.14.0
jinja2>=3.1.0
